requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diffusers==0.31.0
filelock==3.16.1
Flask==3.1.0
Flask-SocketIO==5.4.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==70.2.0
simple-websocket==1.1.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.6.0.dev20241122
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
